feature: добавить эндпоинты получения опций для фильтров по полям исходов заболеваний, результатов диспансеризации, результатов обращения, способов оплаты и других

+++ Infrastructure/Database/DbEntities/References/PaymentMethodDbEntity.cs
@@ -2,7 +2,7 @@
 
 public sealed class PaymentMethodDbEntity
 {
-    public int Uid { get; set; }
+    public short Uid { get; set; }
 
     public string PaymentMethodName { get; set; } = string.Empty;
 }
